qa-lab: expand Matrix live scenario coverage
